requirements.txt
mysql-connector-python
langchain
langchain_community
dotenv
langchain_google_genai
langgraph
langgraph
faiss-cpu
sentence-transformers
langchain_core
langchain_groq
